Moved processing logic
