+++ components/CameraView.tsx
@@ -1,47 +1,24 @@
 import React, { useEffect, useState } from "react";
 import { StyleSheet, View, Text } from "react-native";
 import { Camera } from "expo-camera";
-import * as tf from "@tensorflow/tfjs";
 import "@tensorflow/tfjs-react-native";
-import * as cocoSsd from "@tensorflow-models/coco-ssd";
-import * as FileSystem from "expo-file-system";
 
 import PhotoButton from "../components/PhotoButton";
-import { base64ImageToTensor } from "../util/image";
+export interface CameraViewProps {
+	processImage: (uri: string) => Promise<void>;
+}
 
-const CameraView = (): JSX.Element => {
+const CameraView = ({ processImage }: CameraViewProps): JSX.Element => {
 	const [ref, setRef] = useState<Camera | null>();
 	const [lastImageURI, setLastImageURI] = useState<string>();
-	const [objectDetectionPreidictions, setObjectDetectionPreidictions] =
-		useState<cocoSsd.DetectedObject[]>();
-	const [status, setStatus] = useState<string>("Waiting for image");
-	const [objDetectModel, setObjDetectModel] = useState<cocoSsd.ObjectDetection>();
 
 	const onPressPhotoButton = async (): Promise<void> => {
-		try {
-			const picture = await ref?.takePictureAsync();
-			if (!picture?.uri) {
-				throw new Error("No picture");
-			}
-			setLastImageURI(picture?.uri);
-
-			setStatus("Reading...");
-			const buffer = await FileSystem.readAsStringAsync(picture.uri, {
-				encoding: FileSystem.EncodingType.Base64,
-			});
-			setStatus("Processing...");
-			const imageTensor: tf.Tensor3D = base64ImageToTensor(buffer);
-
-			setStatus("Detecting objects...");
-			const predictions = await objDetectModel?.detect(imageTensor);
-			setObjectDetectionPreidictions(predictions);
-			setStatus(`objects: ${JSON.stringify(objectDetectionPreidictions)}`);
-		} catch (e) {
-			if (e instanceof Error) {
-				setStatus(`Error: ${e.message}`);
-			}
-			throw e;
+		const picture = await ref?.takePictureAsync();
+		if (!picture?.uri) {
+			throw new Error("No picture");
 		}
+		setLastImageURI(picture?.uri);
+		await processImage(picture?.uri);
 	};
 
 	useEffect((): void => {
@@ -55,17 +32,6 @@ const CameraView = (): JSX.Element => {
 					}
 					break;
 				}
-			}
-
-			setStatus("Waiting for tensorflow to be ready...");
-			await tf.ready();
-
-			if (!objDetectModel) {
-				setStatus("Loading object detection model...");
-				const objectDetection = await cocoSsd.load();
-				setObjDetectModel(objectDetection);
-
-				setStatus("Waiting for image");
 			}
 		})();
 	}, []);
@@ -82,7 +48,6 @@ const CameraView = (): JSX.Element => {
 			<PhotoButton onPress={onPressPhotoButton} color="#f4e1e6" />
 			{/* FIXME: temporary, remove when we have actual UI */}
 			<Text>{lastImageURI}</Text>
-			<Text>{status}</Text>
 		</View>
 	);
 };
